+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/games/GameEmulatorRepresentation.java
@@ -66,6 +66,10 @@ public class GameEmulatorRepresentation {
     return emulatorType.isFpEmulator();
   }
 
+  public boolean isFxEmulator() {
+    return emulatorType.isFxEmulator();
+  }
+
   public boolean isVpxEmulator() {
     return emulatorType.isVpxEmulator();
   }