+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/recorder/RecorderController.java
@@ -351,7 +351,7 @@ public class RecorderController extends BaseTableController<GameRepresentation, 
             refreshScreens();
           });
 
-          Thread.sleep(500);
+          Thread.sleep(1000);
         }
       }
       catch (Exception e) {